chore: split geo and zset families
