+++ src/server/zset_family_test.cc
@@ -8,6 +8,7 @@
 #include "base/logging.h"
 #include "facade/facade_test.h"
 #include "server/command_registry.h"
+#include "server/geo_family.h"
 #include "server/test_utils.h"
 
 using namespace testing;
